# swagger.json
{
    "swagger": "2.0",
    "info": {
    "version": "1.0.0", 
    "title": "SuzErrand APIs",
    "description": "Errand Booking Application API",
    "license": {
    "name": "MIT",
    "url": "https://opensource.org/licenses/MIT"
    }
    },
    "host": "localhost:3000/api",
    "basePath": "/",
    "tags": [
    {
    "name": "Users",
    "description": "API for users in the system"
    }
    ],
    "schemes": [
    "http"
    ],
    "consumes": [
    "application/json"
    ],
    "produces": [
    "application/json"
    ],
    "definitions" : {
        "booking" : {
            "type" : "object",
        "properties" : {
            "BookerName": {
                "type" : "string"
            },
            "BookerPhone": {
                "type" : "string"
            },
            "SenderName": {
                "type" : "string"
            },
            "SenderPhone": {
                "type" : "string"
            },
            "RecipientName": {
                "type" : "string"
            },
            "RecipientPhone": {
                "type" : "string"
            },
            "PickupAddress": {
                "type" : "string"
            },
            "DeliveryAddress": {
                "type" : "string"
            },
            "DeliveryDate": {
                "type" : "string"
            },
            "PickupDate": {
                "type" : "string"
            },
            "PaymentMethod": {
                "type" : "string"
            },
            "Payer": {
                "type" : "string"
            },
            "Prioritylevel": {
                "type" : "string"
            }
        }
    },
        "bookings" : {
            "type" : "object",
        "additionalProperties" : {
            "$ref": "#/definitions/booking"
    }
},
        "_id" : {
            "type": "object",
            "properties" : {
                "_id" : {
                    "type" : "string"
                }
            }
        }
},
    "paths" : {
            "/bookings": {
              "get": {
                "description": "Returns all bookings made",
                "responses": {
                  "200": {          
                    "description": "A list of bookings.",
                    "schema": {
                        "type": "array",
                        "items": {
                          "$ref": "#/definitions/bookings"
                        }
                      }
                  }
                }
            }
        },
                "/book": {
                    "parameters" : [
                    {
                     "name" : "body",
                        "in" : "body",
                        "required" : "true",
                        "schema" : {
                        "$ref": "#/definitions/booking"
                     }
                    }
                ],
                "post": {
                    "description": "Returns all orders made",
                     "responses": {
                        "200": {          
                            "description": "Id of created booking.",
                            "schema" : {
                                "$ref" : "#/definitions/_id"
                            }
                        }
                    }
                 }
        },
            
             "/bookings/{id}": {
                "parameters" : [{
                "name" : "id",
                "in" : "path",
                "required" : "true",
                "type" : "string"
            }],
            "get": {
                "description": "Returns the requested booking",
                "responses": {
                    "200": {          
                      "description": "Ok",
                      "schema": {
                        "$ref": "#/definitions/booking"
                      }
                    }
                }
            },
                "put": {
                    "parameters" : [
                        {
                        "name" : "id",
                        "in" : "path",
                        "required" : "true",
                        "type" : "string"
                    },
                    {
                        "name" : "body",
                        "in" : "body",
                        "required" : "true",
                        "schema" : {
                            "$ref": "#/definitions/booking"
                        }
                    }
                ],
                    "description": "Returns all orders made",
                     "responses": {
                        "204": {          
                         "description": "Update."
                    }
                }
            },
            "delete": {
                "parameters" : [
                    {
                    "name" : "id",
                    "in" : "path",
                    "required" : "true",
                    "type" : "string"
                }],
                "description": "Deletes the booking in the request",
                 "responses": {
                    "204": {          
                     "description": "Delete"
                    }
                }
            }

        }
    }
}
  